refactor - code composition
