+++ src/pages/feed.tsx
@@ -1,9 +1,9 @@
-import { Box, Fab } from "../utilities/material-ui/material-components";
+import { Box } from "../utilities/material-ui/material-components";
 import { DiscussionCardsSection } from "../components/DiscussionCardsSection";
 import { Sidenav } from "../components/Sidenav";
 import { TrendingSection } from "../components/TrendingSection";
-import { AddIcon } from "../utilities/material-ui/material-icons";
 import { useState } from "react";
+import { AddPostMobile } from "../components/AddPostMobile";
 
 export function Feed() {
   const [open, setOpen] = useState(false);
@@ -23,19 +23,7 @@ export function Feed() {
       <Sidenav handleOpen={handleOpen} open={open} setOpen={setOpen} />
       <DiscussionCardsSection title="Latest Posts" />
       <TrendingSection />
-      <Fab
-        color="primary"
-        aria-label="add"
-        sx={{
-          position: "fixed",
-          bottom: 24,
-          right: 24,
-          display: { xs: "block", md: "none" },
-        }}
-        onClick={handleOpen}
-      >
-        <AddIcon />
-      </Fab>
+      <AddPostMobile handleOpen={handleOpen} />
     </Box>
   );
 }
